Move `base` model to `base.py` from `core.py`
